requirements.txt
asgiref==3.5.2
Django==4.1.3
django-rest-framework==0.1.0
djangorestframework==3.14.0
Pillow==9.3.0
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6
